+++ app/src/main/java/com/example/movielistapplication/Database/DAOS/MovieDao.java
@@ -1,4 +1,4 @@
-package com.example.movielistapplication.Database;
+package com.example.movielistapplication.Database.DAOS;
 
 import androidx.lifecycle.LiveData;
 import androidx.room.Dao;
@@ -6,6 +6,8 @@ import androidx.room.Delete;
 import androidx.room.Insert;
 import androidx.room.OnConflictStrategy;
 import androidx.room.Query;
+
+import com.example.movielistapplication.Database.MovieListDatabase;
 import com.example.movielistapplication.Database.entities.Movie;
 import java.util.List;
