+++ app/src/main/java/com/example/movielistapplication/Database/MovieListRepository.java
@@ -4,6 +4,8 @@ import android.util.Log;
 
 import androidx.lifecycle.LiveData;
 
+import com.example.movielistapplication.Database.DAOS.MovieDao;
+import com.example.movielistapplication.Database.DAOS.UserDao;
 import com.example.movielistapplication.Database.entities.Movie;
 import com.example.movielistapplication.Database.entities.User;
