+++ app/src/main/res/layout/movie_item.xml
@@ -41,9 +41,10 @@
         <ImageView
             android:id="@+id/movieImageViewForMoviePoster"
             android:layout_width="120dp"
-            android:layout_height="150dp"
-            android:layout_marginStart="15dp"
-            android:layout_marginEnd="15dp"
+            android:layout_height="140dp"
+            android:layout_marginTop="5dp"
+            android:layout_marginStart="5dp"
+            android:layout_marginEnd="5dp"
             android:background="@color/white"
             app:srcCompat="@drawable/baseline_person_24"
             app:layout_constraintEnd_toEndOf="@id/guideLineForMovieItem"
